requirements.txt
#python=3.8.20
absl-py==2.3.1

astunparse==1.6.3

bcrypt==5.0.0
blinker==1.8.2
cachetools==5.5.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8

contourpy==1.1.1
cycler==0.12.1

Flask==3.0.3
flatbuffers==25.9.23
fonttools==4.57.0
gast==0.4.0
google-auth==2.43.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.11.0
idna==3.11

importlib_resources==6.4.5

itsdangerous==2.2.0

Jinja2==3.1.6
joblib==1.4.2

keras==2.13.1
kiwisolver==1.4.7
libclang==18.1.1
lightgbm==4.6.0
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5

mysql-connector-python==9.0.0
narwhals==1.42.1

numpy==1.24.3
oauthlib==3.3.1
opt_einsum==3.4.0

pandas==2.0.3

pillow==10.4.0

plotly==6.5.0

protobuf==4.25.8

pyasn1==0.6.1
pyasn1_modules==0.4.2

pyparsing==3.1.4

pytz==2025.2
pywin32==306

requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2

tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0

tzdata==2025.2
urllib3==2.2.3

Werkzeug==3.0.6
wrapt==2.0.1
xgboost==2.1.4
zipp==3.20.2
